Fixed bugs and added gdt_time_ns
Added gdt_time_ns which returns a time in nano seconds to the highest
precision available.
Added a call to gdt_hook_hidden on Android when hiding the app, was
never called before.
Setting the log threshold on iOS is bugged, changed so that all log
messages are seen by default as a work around.